Explicitly use json parser for .json files
Alongside the discussion in the prettier issues, json files are not always parsed as JSON but rather JSONC or JSON5. Here, prettier doesn't render files according to their file extension.
We, on the other hand, work that way so it's necessary to explicitly set the parser to avoid trailing commas in JSON files.